JobHost.CallAsync can now handle Task.WhenAll.
The Task.WhenAll (and Task.WhenAny) methods return an internal or
private class representing the wrapped tasks at hand. This causes the
CallAsync method to fail, because it attempts to retrieve the generic
parameters when there are none.

Support for WhenAny is not included as this returns a nested task, and
is explicitly tested for and forbidden.

Fixes #660.