Add Rule-based binding extensibility. Also includes updates to EventHub to take advantage of it.
The net goal here is:
 1. extension authors just pick a "rule" and provide a Attribute --> object conversion funciton.
 2. attributes are automatically "resolved" for {} and %% tokens when passed to the rule. This means most extensions no longer need to worry about calling name-resolver or binding contracts.
 3. the "rules" are fully implemented BindingProviders.

So extensions just pick a rule and usually don't need to implement a IBindingProvider, IBinding, IValueProvider, IValueBinder, IArgumentBinding, and they don't need to use INameResolver, BindingContext, ValueBindingContext, AmbientBindingContext, ParameterDescriptor. This also provides default behavior for invokeStrings, invoke/replay, and dashboard support.

The goal is that the "easiest path" for extension authors will lead them to prescribed patterns that
a) are consistent with the core bindings especially in the corner cases;
b) naturally work with Functions and can be called from a variety of Function-supported languages beyond just C#.