Allow SDK to run purely off just a SAS connection string to a container (#1425)
This allows the SDK to run without a full storage account. 
Fix https://github.com/Azure/azure-webjobs-sdk-script/issues/2209

It adds a JobHostInternalStorageConfiguration  object that can speciy a SAS connection to a container. 
We may add other SAS connections over time (ie, we may need one for the shared queue). 
The SDK discovers this via environment variables. 
The new config is exposes publicly so that extensions (like timer) can pull and use it too. 

It includes a strong end-2-end test that creates the SAS connection string and invokes the SDK and calls functions
that excerise the sas container.